consolidated source into simulation.c to remove need for link time optimization
