+++ src/openmp-threading/xs_kernel.c
@@ -1,255 +1,2 @@
 #include "rsbench.h"
 
-// This function uses a combination of the Abrarov Approximation
-// and the QUICK_W three term asymptotic expansion.
-// Only expected to use Abrarov ~0.5% of the time.
-RSComplex fast_nuclear_W( RSComplex Z )
-{
-	// Abrarov 
-	if( c_abs(Z) < 6.0 )
-	{
-		// Precomputed parts for speeding things up
-		// (N = 10, Tm = 12.0)
-		RSComplex prefactor = {0, 8.124330e+01};
-		double an[10] = {
-			2.758402e-01,
-			2.245740e-01,
-			1.594149e-01,
-			9.866577e-02,
-			5.324414e-02,
-			2.505215e-02,
-			1.027747e-02,
-			3.676164e-03,
-			1.146494e-03,
-			3.117570e-04
-		};
-		double neg_1n[10] = {
-			-1.0,
-			1.0,
-			-1.0,
-			1.0,
-			-1.0,
-			1.0,
-			-1.0,
-			1.0,
-			-1.0,
-			1.0
-		};
-
-		double denominator_left[10] = {
-			9.869604e+00,
-			3.947842e+01,
-			8.882644e+01,
-			1.579137e+02,
-			2.467401e+02,
-			3.553058e+02,
-			4.836106e+02,
-			6.316547e+02,
-			7.994380e+02,
-			9.869604e+02
-		};
-
-		RSComplex t1 = {0, 12};
-		RSComplex t2 = {12, 0};
-		RSComplex i = {0,1};
-		RSComplex one = {1, 0};
-		RSComplex W = c_div(c_mul(i, ( c_sub(one, fast_cexp(c_mul(t1, Z))) )) , c_mul(t2, Z));
-		RSComplex sum = {0,0};
-		for( int n = 0; n < 10; n++ )
-		{
-			RSComplex t3 = {neg_1n[n], 0};
-			RSComplex top = c_sub(c_mul(t3, fast_cexp(c_mul(t1, Z))), one);
-			RSComplex t4 = {denominator_left[n], 0};
-			RSComplex t5 = {144, 0};
-			RSComplex bot = c_sub(t4, c_mul(t5,c_mul(Z,Z)));
-			RSComplex t6 = {an[n], 0};
-			sum = c_add(sum, c_mul(t6, c_div(top,bot)));
-		}
-		W = c_add(W, c_mul(prefactor, c_mul(Z, sum)));
-		return W;
-	}
-	else
-	{
-		// QUICK_2 3 Term Asymptotic Expansion (Accurate to O(1e-6)).
-		// Pre-computed parameters
-		RSComplex a = {0.512424224754768462984202823134979415014943561548661637413182,0};
-		RSComplex b = {0.275255128608410950901357962647054304017026259671664935783653, 0};
-		RSComplex c = {0.051765358792987823963876628425793170829107067780337219430904, 0};
-		RSComplex d = {2.724744871391589049098642037352945695982973740328335064216346, 0};
-
-		RSComplex i = {0,1};
-		RSComplex Z2 = c_mul(Z, Z);
-		// Three Term Asymptotic Expansion
-		RSComplex W = c_mul(c_mul(Z,i), (c_add(c_div(a,(c_sub(Z2, b))) , c_div(c,(c_sub(Z2, d))))));
-
-		return W;
-	}
-}
-
-void calculate_macro_xs( double * macro_xs, int mat, double E, Input input, SimulationData data ) 
-{
-	// zero out macro vector
-	for( int i = 0; i < 4; i++ )
-		macro_xs[i] = 0;
-
-	// for nuclide in mat
-	for( int i = 0; i < data.num_nucs[mat]; i++ )
-	{
-		double micro_xs[4];
-		int nuc = data.mats[mat * data.max_num_nucs + i];
-
-		if( input.doppler == 1 )
-			calculate_micro_xs_doppler( micro_xs, nuc, E, input, data );
-		else
-			calculate_micro_xs( micro_xs, nuc, E, input, data);
-
-		for( int j = 0; j < 4; j++ )
-		{
-			macro_xs[j] += micro_xs[j] * data.concs[mat * data.max_num_nucs + i];
-		}
-		// Debug
-		/*
-		printf("E = %.2lf, mat = %d, macro_xs[0] = %.2lf, macro_xs[1] = %.2lf, macro_xs[2] = %.2lf, macro_xs[3] = %.2lf\n",
-		E, mat, macro_xs[0], macro_xs[1], macro_xs[2], macro_xs[3] );
-		*/
-	}
-
-	// Debug
-	/*
-	printf("E = %.2lf, mat = %d, macro_xs[0] = %.2lf, macro_xs[1] = %.2lf, macro_xs[2] = %.2lf, macro_xs[3] = %.2lf\n",
-	E, mat, macro_xs[0], macro_xs[1], macro_xs[2], macro_xs[3] );
-	*/
-	
-}
-
-// No Temperature dependence (i.e., 0K evaluation)
-void calculate_micro_xs( double * micro_xs, int nuc, double E, Input input, SimulationData data)
-{
-	// MicroScopic XS's to Calculate
-	double sigT;
-	double sigA;
-	double sigF;
-	double sigE;
-
-	// Calculate Window Index
-	double spacing = 1.0 / data.n_windows[nuc];
-	int window = (int) ( E / spacing );
-	if( window == data.n_windows[nuc] )
-		window--;
-
-	// Calculate sigTfactors
-	RSComplex sigTfactors[4]; // Of length input.numL, which is always 4
-	calculate_sig_T(nuc, E, input, data, sigTfactors );
-
-	// Calculate contributions from window "background" (i.e., poles outside window (pre-calculated)
-	Window w = data.windows[nuc * data.max_num_windows + window];
-	sigT = E * w.T;
-	sigA = E * w.A;
-	sigF = E * w.F;
-
-	// Loop over Poles within window, add contributions
-	for( int i = w.start; i < w.end; i++ )
-	{
-		RSComplex PSIIKI;
-		RSComplex CDUM;
-		Pole pole = data.poles[nuc * data.max_num_poles + i];
-		RSComplex t1 = {0, 1};
-		RSComplex t2 = {sqrt(E), 0 };
-		PSIIKI = c_div( t1 , c_sub(pole.MP_EA,t2) );
-		RSComplex E_c = {E, 0};
-		CDUM = c_div(PSIIKI, E_c);
-		sigT += (c_mul(pole.MP_RT, c_mul(CDUM, sigTfactors[pole.l_value])) ).r;
-		sigA += (c_mul( pole.MP_RA, CDUM)).r;
-		sigF += (c_mul(pole.MP_RF, CDUM)).r;
-	}
-
-	sigE = sigT - sigA;
-
-	micro_xs[0] = sigT;
-	micro_xs[1] = sigA;
-	micro_xs[2] = sigF;
-	micro_xs[3] = sigE;
-}
-
-// Temperature Dependent Variation of Kernel
-// (This involves using the Complex Faddeeva function to
-// Doppler broaden the poles within the window)
-void calculate_micro_xs_doppler( double * micro_xs, int nuc, double E, Input input, SimulationData data )
-{
-	// MicroScopic XS's to Calculate
-	double sigT;
-	double sigA;
-	double sigF;
-	double sigE;
-
-	// Calculate Window Index
-	double spacing = 1.0 / data.n_windows[nuc];
-	int window = (int) ( E / spacing );
-	if( window == data.n_windows[nuc] )
-		window--;
-
-	//printf("spacing = %lf\n", spacing);
-	//printf("window = %d\n", window);
-	// Calculate sigTfactors
-	RSComplex sigTfactors[4]; // Of length input.numL, which is always 4
-	calculate_sig_T(nuc, E, input, data, sigTfactors );
-
-	// Calculate contributions from window "background" (i.e., poles outside window (pre-calculated)
-	Window w = data.windows[nuc * data.max_num_windows + window];
-	sigT = E * w.T;
-	sigA = E * w.A;
-	sigF = E * w.F;
-
-	double dopp = 0.5;
-
-	// Loop over Poles within window, add contributions
-	for( int i = w.start; i < w.end; i++ )
-	{
-		Pole pole = data.poles[nuc * data.max_num_poles + i];
-		//printf("pole: n = %d  p_idx = %d\n", nuc, i);
-
-		// Prep Z
-		RSComplex E_c = {E, 0};
-		RSComplex dopp_c = {dopp, 0};
-		RSComplex Z = c_mul(c_sub(E_c, pole.MP_EA), dopp_c);
-
-		// Evaluate Fadeeva Function
-		RSComplex faddeeva = fast_nuclear_W( Z );
-		//RSComplex faddeeva = {1,1};
-		//printf("real fad = %lf\n", faddeeva.r);
-
-		// Update W
-		sigT += (c_mul( pole.MP_RT, c_mul(faddeeva, sigTfactors[pole.l_value]) )).r;
-		sigA += (c_mul( pole.MP_RA , faddeeva)).r;
-		sigF += (c_mul( pole.MP_RF , faddeeva)).r;
-	}
-
-	sigE = sigT - sigA;
-
-	micro_xs[0] = sigT;
-	micro_xs[1] = sigA;
-	micro_xs[2] = sigF;
-	micro_xs[3] = sigE;
-}
-
-void calculate_sig_T( int nuc, double E, Input input, SimulationData data, RSComplex * sigTfactors )
-{
-	double phi;
-
-	for( int i = 0; i < 4; i++ )
-	{
-		phi = data.pseudo_K0RS[nuc * input.numL + i] * sqrt(E);
-
-		if( i == 1 )
-			phi -= - atan( phi );
-		else if( i == 2 )
-			phi -= atan( 3.0 * phi / (3.0 - phi*phi));
-		else if( i == 3 )
-			phi -= atan(phi*(15.0-phi*phi)/(15.0-6.0*phi*phi));
-
-		phi *= 2.0;
-
-		sigTfactors[i].r = cos(phi);
-		sigTfactors[i].i = -sin(phi);
-	}
-}
